# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: 受験マップ API
  version: "1"
servers:
  - url: https://juken-map.com
security:
  - sessionCookie: []
paths:
...
  /api/admin/users/{id}:
    delete:
      operationId: deleteUser
      summary: 利用者を消す。確認のため、本人のメールアドレスが一致しないと消さない
      parameters:
        - $ref: "#/components/parameters/AdminUserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, minLength: 1, maxLength: 191, description: 大文字小文字と前後の空白は無視して比べる }
      responses:
        "200":
          description: 消した。ぶら下がる行は外部キーの CASCADE で一緒に消える
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDeleteResult" }
        "400":
          description: >-
            入力チェックで弾いた（ValidationError の形：文言・コード・項目）か、
            メールアドレスが一致しない（文言だけ）。どちらも error は必ずある
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/AdminUserProtected" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ---- マスター編集（/admin/masters、JUK-78）。変更はすべて構造化ログ「admin master change」に残す ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-jm_session
      description: ログイン（Go の /api/auth/*、JUK-115）が発行するセッション Cookie。中身の無い 256 ビットの乱数のトークン
    cronBearer:
      type: http
      scheme: bearer
      description: DAILY_NOTIFICATION_SECRET
    simBearer:
      type: http
      scheme: bearer
      description: SIMULATION_SECRET（cron とは別のトークン）
  parameters:
    AdminUserID:
      name: id
      in: path
      required: true
      description: 利用者の ID（Better Auth の user.id）
      schema: { type: string, minLength: 1, maxLength: 191 }
...
  responses:
    AdminForbidden:
      description: >-
        管理者でない（{"error":"Forbidden"}）、2段階認証を通していない（code が TWO_FACTOR_REQUIRED）、
        利用を停止されている、別のサイトからの書き込み
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AdminUserProtected:
      description: 相手はいるが守られている（自分自身・他の管理者・デモアカウント。削除はメールアドレスの無い相手も）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: ログインしていない（またはトークンが違う）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: 見つからない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: 本文が上限（ふだんは 1MiB、CSP の報告は 16KB）を超えた
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
    UnsupportedMediaType:
      description: 受け付けない Content-Type（受けるのは JSON・text/plain・CSP の報告の2種類。本文が無ければ無しでよい）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
    InternalError:
      description: 想定外の失敗。原因はログにだけ残す
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
  schemas:
...
    AdminDeleteResult:
      type: object
      required: [id, email, removed]
      properties:
        id: { type: string }
        email: { type: string, nullable: true }
        removed:
          description: 一緒に消えた行数（記録のためだけ。消すのは外部キーの CASCADE）
          type: object
          required: [studyLogs, studyPlans, textbooks, finalGoals]
          properties:
            studyLogs: { type: integer }
            studyPlans: { type: integer }
            textbooks: { type: integer }
            finalGoals: { type: integer }
...
    Error:
      description: ルートが自分で断るときのエラー。error は画面にそのまま出す文言
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ServerError:
      description: 想定外の失敗・混雑・存在しないパス
      type: object
      required: [error, code, reqId]
      properties:
        error: { type: string }
        code:
          type: string
          # FST_ERR_CTP_* は Fastify の code をそのまま使っている（Node の error-handling.ts）
          # FST_ERR_CTP_INVALID_CONTENT_LENGTH は本文が UTF-8 として不正で、読んだ長さが Content-Length と合わないとき（400）
          enum:
            - INTERNAL_ERROR
            - OVERLOADED
            - NOT_FOUND
            - FST_ERR_CTP_BODY_TOO_LARGE
            - FST_ERR_CTP_INVALID_MEDIA_TYPE
            - FST_ERR_CTP_INVALID_CONTENT_LENGTH
            # 利用者単位の回数制限を超えた（429、Go の internal/httpx/user_rate_limit.go）
            - TOO_MANY_REQUESTS
          x-enum-varnames:
            - ServerErrorCodeInternal
            - ServerErrorCodeOverloaded
            - ServerErrorCodeNotFound
            - ServerErrorCodeBodyTooLarge
            - ServerErrorCodeInvalidMediaType
            - ServerErrorCodeInvalidContentLength
            - ServerErrorCodeTooManyRequests
        reqId: { type: string }
...